LayoutTests/platform/win/plugins/plugin-document-back-forward-expected.txt

ALERT: Plugin Loaded!
ALERT: Plugin Loaded!


--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
